+++ bitswap/options.go
@@ -71,12 +71,12 @@ func WithTaskComparator(comparator server.TaskComparator) Option {
 	return Option{server.WithTaskComparator(comparator)}
 }
 
-// WithReplaceHasWithBlockMaxSize sets the maximum size of a block in bytes up
-// to which we will replace a want-have with a want-block. Setting a size of 0
-// disables this want-have replacement and means that block sizes are not read
-// for want-have requests.
-func WithReplaceHasWithBlockMaxSize(maxSize int) Option {
-	return Option{server.WithReplaceHasWithBlockMaxSize(maxSize)}
+// WithWantHaveReplaceSize sets the maximum size of a block in bytes up to
+// which the bitswap server will replace a WantHave with a WantBlock response.
+// Setting this to 0 disables this WantHave replacement and means that block
+// sizes are not read when processing WantHave requests.
+func WithWantHaveReplaceSize(size int) Option {
+	return Option{server.WithWantHaveReplaceSize(size)}
 }
 
 func ProviderSearchDelay(newProvSearchDelay time.Duration) Option {